+++ src/Query/helpers.js
@@ -1,15 +1,9 @@
 // @flow
 
-import { uniq } from 'rambdax'
-
 import zip from '../utils/fp/zip'
 
-import type { QueryDescription } from '../QueryDescription'
 import type { Associations, AssociationInfo } from '../Model'
 import type { TableName } from '../Schema'
-
-export const getSecondaryTables: QueryDescription => TableName<any>[] = description =>
-  uniq(description.join.map(join => join.table))
 
 export const getAssociations: (
   TableName<any>[],